Add fallback images to exercises without icons
