+++ app/css/pages/journey.css
@@ -14,7 +14,7 @@
             @apply absolute;
             left: 0;
             right: 0;
-            bottom: 0;
+            top: 0;
 
             & .bottom {
                 @apply absolute;
