+++ app/javascript/components/common/imageErrorHandler.tsx
@@ -1,0 +1,12 @@
+const errorIcon = require(`../../images/graphics/missing-exercise.svg`)
+
+export const imageErrorHandler = (
+  e: React.SyntheticEvent<HTMLImageElement, Event>
+) => {
+  const el = e.target as HTMLImageElement
+  if ((el.src = errorIcon)) {
+    return
+  }
+  el.onerror = null
+  el.src = errorIcon
+}
