+++ app/javascript/components/common/index.ts
@@ -24,3 +24,4 @@ export { SingleSelect } from './SingleSelect'
 export { MultipleSelect } from './MultipleSelect'
 export { SiteUpdatesList } from './SiteUpdatesList'
 export { ProgressGraph } from './ProgressGraph'
+export { imageErrorHandler } from './imageErrorHandler'
